pyproject.toml
[tool.poetry]
name = "freelance-parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
telethon = "^1.33.1"
pytz = "^2024.1"
pytelegrambotapi = "^4.15.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
